- Fix Inconsistency Model Personality Insights - Fix Aggregate Exception - Fix Service Name in Personality Insights Service
