+++ src/IBM.WatsonDeveloperCloud.PersonalityInsights/v3/Model/Languages.cs
@@ -1,0 +1,33 @@
+﻿/**
+* Copyright 2017 IBM Corp. All Rights Reserved.
+*
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+*
+*      http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*
+*/
+
+using IBM.WatsonDeveloperCloud.Util.Attributes;
+
+namespace IBM.WatsonDeveloperCloud.PersonalityInsights.v3.Model
+{
+    public enum Languages
+    {
+        [Description("ar")]
+        ARABIC,
+        [Description("en")]
+        ENGLISH,
+        [Description("es")]
+        SPANISH,
+        [Description("ja")]
+        JAPANESE
+    }
+}
