+++ src/IBM.WatsonDeveloperCloud.PersonalityInsights/v3/Model/TraitTreeNode.cs
@@ -1,0 +1,43 @@
+﻿/**
+* Copyright 2017 IBM Corp. All Rights Reserved.
+*
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+*
+*      http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*
+*/
+
+using System.Collections.Generic;
+using Newtonsoft.Json;
+
+namespace IBM.WatsonDeveloperCloud.PersonalityInsights.v3.Model
+{
+    public class TraitTreeNode
+    {
+        [JsonProperty("trait_id")]
+        public string TraitId { get; set; }
+
+        [JsonProperty("name")]
+        public string Name { get; set; }
+
+        [JsonProperty("category")]
+        public string Category { get; set; }
+
+        [JsonProperty("percentile")]
+        public double Percentile { get; set; }
+
+        [JsonProperty("raw_score")]
+        public double RawScore { get; set; }
+
+        [JsonProperty("children")]
+        public List<TraitTreeNode> Children { get; set; }
+    }
+}
