--- src/IBM.WatsonDeveloperCloud.PersonalityInsights/v3/Models/Warning.cs
@@ -1,30 +1,0 @@
-﻿/**
-* Copyright 2017 IBM Corp. All Rights Reserved.
-*
-* Licensed under the Apache License, Version 2.0 (the "License");
-* you may not use this file except in compliance with the License.
-* You may obtain a copy of the License at
-*
-*      http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific language governing permissions and
-* limitations under the License.
-*
-*/
-
-using Newtonsoft.Json;
-
-namespace IBM.WatsonDeveloperCloud.PersonalityInsights.v3.Models
-{
-    public class Warning
-    {
-        [JsonProperty("warning_id")]
-        public string WarningId { get; set; }
-
-        [JsonProperty("message")]
-        public string Message { get; set; }
-    }
-}
