+++ test/IBM.WatsonDeveloperCloud.PersonalityInsights.IntegrationTests/PersonalityInsightsServiceIntegrationTest.cs
@@ -19,7 +19,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using IBM.WatsonDeveloperCloud.PersonalityInsights.v3.Models;
+using IBM.WatsonDeveloperCloud.PersonalityInsights.v3.Model;
 
 namespace IBM.WatsonDeveloperCloud.PersonalityInsights.IntegrationTests
 {
